Set username using adfs email
