+++ ckanext/adfs/plugin.py
@@ -10,17 +10,20 @@ import base64
 from validation import validate_saml
 from metadata import get_certificates, get_federation_metadata, get_wsfed
 from extract import get_user_info
-
+try:
+    from ckan.common import config
+except ImportError:
+    from pylons import config
 
 log = logging.getLogger(__name__)
 
 
 # Some awful XML munging.
 WSFED_ENDPOINT = ''
-WTREALM = pylons.config['adfs_wtrealm']
-METADATA = get_federation_metadata(pylons.config['adfs_metadata_url'])
+WTREALM = config['adfs_wtrealm']
+METADATA = get_federation_metadata(config['adfs_metadata_url'])
 WSFED_ENDPOINT = get_wsfed(METADATA)
-AUTH_URL_TEMPLATE = pylons.config.get('adfs_url_template','{}?wa=wsignin1.0&wreq=xml&wtrealm={}')
+AUTH_URL_TEMPLATE = config.get('adfs_url_template','{}?wa=wsignin1.0&wreq=xml&wtrealm={}')
 
 
 if not (WSFED_ENDPOINT):
@@ -48,11 +51,11 @@ class ADFSPlugin(plugins.SingletonPlugin):
     plugins.implements(plugins.IRoutes)
     plugins.implements(plugins.IAuthenticator)
 
-    def update_config(self, config):
+    def update_config(self, config_):
         """
         Add our templates to CKAN's search path
         """
-        toolkit.add_template_directory(config, 'templates')
+        toolkit.add_template_directory(config_, 'templates')
 
     def get_helpers(self):
         return dict(is_adfs_user=is_adfs_user,
@@ -69,7 +72,7 @@ class ADFSPlugin(plugins.SingletonPlugin):
         # Route requests for our WAAD redirect URI to a custom controller.
         map.connect(
             'adfs_redirect_uri', '/adfs/signin/',
-            controller='ckanext.adfs.plugin:ADFSRedirectController',
+            controller='ckanext.adfs.controller:ADFSRedirectController',
             action='login')
         return map
 
@@ -108,81 +111,9 @@ class ADFSPlugin(plugins.SingletonPlugin):
                 del pylons.session[key]
             pylons.session.save()
 
-
     def abort(self, status_code, detail, headers, comment):
         """
         Called on abort.  This allows aborts due to authorization issues
         to be overriden.
         """
         return (status_code, detail, headers, comment)
-
-
-def _get_user(name):
-    """
-    Return the CKAN user with the given user name, or None.
-    """
-    try:
-        return toolkit.get_action('user_show')(data_dict = {'id': name})
-    except toolkit.ObjectNotFound:
-        return None
-
-
-class FileNotFoundException(Exception):
-    pass
-
-
-class ADFSRedirectController(toolkit.BaseController):
-    """
-    A custom home controller for receiving ADFS authorization responses.
-    """
-
-    def login(self):
-        """
-        Handle eggsmell request from the ADFS redirect_uri.
-        """
-        try:
-            eggsmell = pylons.request.POST.get('wresult')
-            if not eggsmell:
-                request_data = dict(pylons.request.POST)
-                eggsmell = base64.decodestring(request_data['SAMLResponse'])
-        except:
-            log.info('ADFS eggsmell')
-            log.info(dict(pylons.request.POST))
-        # We grab the metadata for each login because due to opaque
-        # bureaucracy and lack of communication the certificates can be
-        # changed. We looked into this and took made the call based upon lack
-        # of user problems and tech being under our control vs the (small
-        # amount of) latency from a network call per login attempt.
-        metadata = get_federation_metadata(pylons.config['adfs_metadata_url'])
-        x509_certificates = get_certificates(metadata)
-        if not validate_saml(eggsmell, x509_certificates):
-            raise ValueError('Invalid signature')
-        username, email, firstname, surname = get_user_info(eggsmell)
-
-        if not email:
-            log.error('Unable to login with ADFS')
-            log.error(eggsmell)
-            raise ValueError('No email returned with ADFS')
-
-        user = _get_user(username)
-        if user:
-            # Existing user
-            log.info('Logging in from ADFS with user: {}'.format(username))
-        else:
-            # New user, so create a record for them.
-            log.info('Creating user from ADFS')
-            log.info('email: {} firstname: {} surname: {}'.format(email,
-                     firstname.encode('utf8'), surname.encode('utf8')))
-            log.info('Generated username: {}'.format(username))
-            # TODO: Add the new user to the NHSEngland group? Check this!
-            user = toolkit.get_action('user_create')(
-                context={'ignore_auth': True},
-                data_dict={'name': username,
-                           'fullname': firstname + ' ' + surname,
-                           'password': str(uuid.uuid4()),
-                           'email': email})
-        pylons.session['adfs-user'] = username
-        pylons.session['adfs-email'] = email
-        pylons.session.save()
-        toolkit.redirect_to(controller='user', action='dashboard', id=email)
-        return
